# perf: disable checking names in data.frame call