Checking names in data.frame call is rather expensive:

``` r
x <- rnorm(1e5)
bench::mark(
  data.frame(x = x),
  data.frame(x = x, check.names = FALSE)
)
#> # A tibble: 2 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                          <bch:> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 data.frame(x = x)                   30.1µs 33.4µs    28719.        0B     37.4
#> 2 data.frame(x = x, check.names = FA… 21.8µs 23.5µs    40706.        0B     40.7
```

<sup>Created on 2025-09-18 with [reprex v2.1.1](https://reprex.tidyverse.org)</sup>